pyproject.toml
[project]
name = "bigquery-sodirchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-cloud-aiplatform>=1.115.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-bigquery-storage>=2.33.1",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "tqdm>=4.67.1",
]
